Handle miscellaneous warnings (#47)
* Handle miscellaneous warnings

* Improve memset usage here

+++ cbase64.h
@@ -120,7 +120,7 @@ unsigned int cbase64_encode_block(const unsigned char* data_in, unsigned int len
             {
                 state_in->step = step_A;
                 state_in->result = result;
-                return codechar - code_out;
+                return uint32_t(codechar - code_out);
             }
             fragment = *datachar++;
             result = (fragment & 0x0fc) >> 2;
@@ -131,7 +131,7 @@ unsigned int cbase64_encode_block(const unsigned char* data_in, unsigned int len
             {
                 state_in->step = step_B;
                 state_in->result = result;
-                return codechar - code_out;
+                return uint32_t(codechar - code_out);
             }
             fragment = *datachar++;
             result |= (fragment & 0x0f0) >> 4;
@@ -142,7 +142,7 @@ unsigned int cbase64_encode_block(const unsigned char* data_in, unsigned int len
             {
                 state_in->step = step_C;
                 state_in->result = result;
-                return codechar - code_out;
+                return uint32_t(codechar - code_out);
             }
             fragment = *datachar++;
             result |= (fragment & 0x0c0) >> 6;
@@ -152,7 +152,7 @@ unsigned int cbase64_encode_block(const unsigned char* data_in, unsigned int len
         }
     }
     // control should not reach here
-    return codechar - code_out;
+    return uint32_t(codechar - code_out);
 }
 
 unsigned int cbase64_decode_block(const char* code_in, unsigned int length_in,
@@ -174,7 +174,7 @@ unsigned int cbase64_decode_block(const char* code_in, unsigned int length_in,
                 {
                     state_in->step = step_A;
                     state_in->result = overwrite;
-                    return datachar - data_out;
+                    return uint32_t(datachar - data_out);
                 }
                 fragment = cbase64__decode_value(*codechar++);
             } while (fragment < 0);
@@ -185,7 +185,7 @@ unsigned int cbase64_decode_block(const char* code_in, unsigned int length_in,
                 {
                     state_in->step = step_B;
                     state_in->result = overwrite;
-                    return datachar - data_out;
+                    return uint32_t(datachar - data_out);
                 }
                 fragment = cbase64__decode_value(*codechar++);
             } while (fragment < 0);
@@ -197,7 +197,7 @@ unsigned int cbase64_decode_block(const char* code_in, unsigned int length_in,
                 {
                     state_in->step = step_C;
                     state_in->result = overwrite;
-                    return datachar - data_out;
+                    return uint32_t(datachar - data_out);
                 }
                 fragment = cbase64__decode_value(*codechar++);
             } while (fragment < 0);
@@ -209,7 +209,7 @@ unsigned int cbase64_decode_block(const char* code_in, unsigned int length_in,
                 {
                     state_in->step = step_D;
                     state_in->result = overwrite;
-                    return datachar - data_out;
+                    return uint32_t(datachar - data_out);
                 }
                 fragment = cbase64__decode_value(*codechar++);
             } while (fragment < 0);
@@ -217,7 +217,7 @@ unsigned int cbase64_decode_block(const char* code_in, unsigned int length_in,
         }
     }
     // control should not reach here
-    return datachar - data_out;
+    return uint32_t(datachar - data_out);
 }
 
 unsigned int cbase64_encode_blockend(char* code_out, cbase64_encodestate* state_in)
@@ -237,7 +237,7 @@ unsigned int cbase64_encode_blockend(char* code_out, cbase64_encodestate* state_
     case step_A:
         break;
     }
-    return codechar - code_out;
+    return uint32_t(codechar - code_out);
 }
 
 #endif // CBASE64_IMPLEMENTATION
